+++ atcg_lib/src/Math/ColorDetail.h
@@ -1,7 +1,7 @@
 #pragma once
 
+#include <Utils/Utils.h>
 #include <Math/Constants.h>
-#include <Math/Utils.h>
 #include <Math/Color.h>
 
 namespace atcg
