+++ atcg_lib/src/Network/TCPClient.cpp
@@ -2,7 +2,7 @@
 
 #include <Core/Assert.h>
 #include <DataStructure/TorchUtils.h>
-#include <Math/Utils.h>
+#include <Utils/Utils.h>
 #include <Network/NetworkUtils.h>
 
 namespace atcg
@@ -76,7 +76,7 @@ std::vector<uint8_t> TCPClient::sendAndWait(const std::vector<uint8_t>& data)
         total_received += received;
     } while(total_received < sizeof(uint32_t));
 
-    uint32_t expected_size = atcg::hton<uint32_t>(message_size);
+    uint32_t expected_size = atcg::Utils::hton<uint32_t>(message_size);
     std::vector<uint8_t> rec_data(expected_size);
     total_received = 0;
     while(total_received < expected_size)
