+++ atcg_lib/src/Utils/Utils.cpp
@@ -1,10 +1,12 @@
-#include <Math/Utils.h>
+#include <Utils/Utils.h>
 
 #include <fstream>
 
 namespace atcg
 {
 
+namespace Utils
+{
 void normalize(const atcg::ref_ptr<Graph>& graph)
 {
     auto vertices = graph->getPositions(atcg::GPU);
@@ -155,13 +157,11 @@ uint64_t ntoh<uint64_t>(uint64_t network)
     return network;
 }
 
-namespace IO
-{
 void dumpBinary(const std::string& path, const torch::Tensor& data)
 {
     auto data_ = data.to(atcg::CPU);
     std::ofstream out(path, std::ios::out | std::ios::binary);
     out.write((const char*)data_.data_ptr(), data_.numel() * data_.element_size());
 }
-}    // namespace IO
+}    // namespace Utils
 }    // namespace atcg
